requirements.txt
ipykernel==6.29.4
pillow==10.3.0
ultralytics==8.3.78
manga-ocr==0.1.14
googletrans==4.0.0-rc1
deep-translator==1.11.4
fastapi
uvicorn==0.30.0
torch==2.6.0
python-bidi==0.6.6
arabic-reshaper==3.0.0
